+++ src/Enums/WebhookEventKind.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Th3JK\Treasurer\Enums;
+
+enum WebhookEventKind: string
+{
+    case PAYMENT_NOTIFICATION = 'payment.notification';
+}