+++ src/TreasurerServiceProvider.php
@@ -54,5 +54,6 @@ class TreasurerServiceProvider extends ServiceProvider
         }
 
         $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
+        $this->loadRoutesFrom(__DIR__.'/../routes/web.php');
     }
 }